Further reducing size of TurnController, now working on divergent change
